Reset tool-guide layout to app root and bump article readability
Rename info/tools/+layout.svelte -> +layout@.svelte so the guide pages
break out of (admin)/+layout.svelte and render without the admin app
sidebar (auth still enforced in (admin)/+layout.ts's load). Make the
guide nav responsive (pills on mobile, left rail from md up) and left
align the layout, letting GuideArticle own the reading measure. Bump
GuideArticle heading/body sizes for readability.
